Clean up table classes
