+++ netbox/ipam/tables/__init__.py
@@ -1,0 +1,4 @@
+from .ip import *
+from .services import *
+from .vlans import *
+from .vrfs import *
